+++ app/src/main/java/de/baumann/browser/view/NinjaWebView.java
@@ -133,6 +133,12 @@ public class NinjaWebView extends WebView implements AlbumController {
         webSettings.setSupportZoom(true);
         webSettings.setBuiltInZoomControls(true);
         webSettings.setDisplayZoomControls(false);
+
+        setVerticalScrollBarEnabled(true);
+        setHorizontalScrollBarEnabled(true);
+        setScrollBarStyle(View.SCROLLBARS_INSIDE_OVERLAY);
+        setScrollbarFadingEnabled(false);
+
         webSettings.setSupportMultipleWindows(true);
         webSettings.setAllowFileAccess(true);
         webSettings.setAllowFileAccessFromFileURLs(true);
@@ -377,16 +383,14 @@ public class NinjaWebView extends WebView implements AlbumController {
                 FloatingActionButton buttonProfile = dialogView.findViewById(R.id.buttonProfile);
                 NinjaWebView.getBrowserController().setProfileIcon(buttonProfile, urlToLoad);
                 FaviconHelper.setFavicon(context, dialogView, url, R.id.menu_icon, R.drawable.icon_image_broken);
-                buttonProfile.setOnClickListener(v -> {
+                buttonProfile.setOnClickListener(v -> NinjaWebView.getBrowserController().showDialogFastToggle(HelperUnit.domain(urlToLoad),urlToLoad, buttonProfile));
+                buttonProfile.setOnLongClickListener(v -> {
                     sp.edit().putString("profile", "profileStandard").apply();
                     NinjaWebView.getBrowserController().setProfileIcon(buttonProfile, urlToLoad);
                     listStandard = new List_standard(context);
                     if (!listStandard.isWhite(urlToLoad)){
                         reload();
                     }
-                });
-                buttonProfile.setOnLongClickListener(v -> {
-                    NinjaWebView.getBrowserController().showDialogFastToggle(HelperUnit.domain(urlToLoad),urlToLoad, buttonProfile);
                     return false;
                 });